Add files via upload
This zip file contains all the code files and images related to the project. The project has done in pure php, html, CSS, JavaScript and AJAX using the SQL language for entering data to the database. Here XAMPP is used as the local web server. MariaDB is used for creating database tables. This is done as an individual project.